flask_files/requirements.txt
boto3 # pre-installed on Lambda, but not Docker image
numpy<2.0.0 # the new numpy has compatibility issues with onnxruntime 1.18.0
pillow>=10.3.0
onnxruntime>=1.18.0
flask>=2.3.2
gunicorn>=20.1.0
